Moved artemis code in node-api into internal package.
None of this needs to be (or should be) exposed publicly to app developers.
