+++ client/rpc/src/main/kotlin/net/corda/client/rpc/CordaRPCClient.kt
@@ -6,8 +6,8 @@ import net.corda.client.rpc.serialization.KryoClientSerializationScheme
 import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.serialization.SerializationDefaults
 import net.corda.core.utilities.NetworkHostAndPort
-import net.corda.nodeapi.ArtemisTcpTransport.Companion.tcpTransport
-import net.corda.nodeapi.ConnectionDirection
+import net.corda.nodeapi.internal.artemis.ArtemisTcpTransport.Companion.tcpTransport
+import net.corda.nodeapi.internal.artemis.ConnectionDirection
 import net.corda.nodeapi.config.SSLConfiguration
 import net.corda.nodeapi.internal.serialization.AMQPClientSerializationScheme
 import net.corda.nodeapi.internal.serialization.KRYO_P2P_CONTEXT
