+++ client/rpc/src/main/kotlin/net/corda/client/rpc/internal/RPCClient.kt
@@ -9,11 +9,11 @@ import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.loggerFor
 import net.corda.core.utilities.minutes
 import net.corda.core.utilities.seconds
-import net.corda.nodeapi.ArtemisTcpTransport.Companion.tcpTransport
-import net.corda.nodeapi.ConnectionDirection
 import net.corda.nodeapi.RPCApi
 import net.corda.nodeapi.RPCException
 import net.corda.nodeapi.config.SSLConfiguration
+import net.corda.nodeapi.internal.artemis.ArtemisTcpTransport.Companion.tcpTransport
+import net.corda.nodeapi.internal.artemis.ConnectionDirection
 import org.apache.activemq.artemis.api.core.SimpleString
 import org.apache.activemq.artemis.api.core.TransportConfiguration
 import org.apache.activemq.artemis.api.core.client.ActiveMQClient
