+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/artemis/ArtemisUtils.kt
@@ -1,6 +1,6 @@
 @file:JvmName("ArtemisUtils")
 
-package net.corda.nodeapi
+package net.corda.nodeapi.internal.artemis
 
 import java.nio.file.FileSystems
 import java.nio.file.Path
