+++ verifier/src/integration-test/kotlin/net/corda/verifier/VerifierDriver.kt
@@ -11,9 +11,9 @@ import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.utilities.NetworkHostAndPort
 import net.corda.core.utilities.loggerFor
 import net.corda.node.services.config.configureDevKeyAndTrustStores
-import net.corda.nodeapi.ArtemisMessagingComponent.Companion.NODE_USER
-import net.corda.nodeapi.ArtemisTcpTransport
-import net.corda.nodeapi.ConnectionDirection
+import net.corda.nodeapi.internal.artemis.ArtemisMessagingComponent.Companion.NODE_USER
+import net.corda.nodeapi.internal.artemis.ArtemisTcpTransport
+import net.corda.nodeapi.internal.artemis.ConnectionDirection
 import net.corda.nodeapi.VerifierApi
 import net.corda.nodeapi.config.NodeSSLConfiguration
 import net.corda.nodeapi.config.SSLConfiguration
